// src/openapi/generated/models/task-application-resident-day.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TaskApplicationDetail } from '../models/task-application-detail';

/**
 * Esquema para aplicaciones de un residente en un día específico
 */
export interface TaskApplicationResidentDay {
  applications: Array<TaskApplicationDetail>;
  bed_name?: (string | null);
  date: string;
  resident_full_name: string;
  resident_id: string;
}